Depend updates.
Censored-Web's future is uncertain, but we'll likely end up using it as an easy 'Make this page now with these arguments' sorta thin.

Allow plugins to be made in Scala 2.10.3.

We'll be using Netty and related libraries (because Netty is already provided by Mojang in the vanilla server).  We'll need to see what version of Netty is loaded on the server before we can do anything with this.